backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
SQLAlchemy
psycopg2-binary
pydantic
pytesseract
Pillow
pdf2image
email-validator
python-multipart
alembic